requirements.txt
matplotlib==3.4.2
torch==1.9.0
numpy==1.21.0
ConfigArgParse==1.5.1
foolbox==3.3.1
pandas==1.3.0
uutils==0.0.3
lmdb==1.2.1
torchvision==0.10.0
Pillow==9.1.0
PyYAML==6.0
scikit_learn==1.0.2
torchsummaryX==1.3.0
